feat(select): multi select component with checkboxes for picking options
A multi select component that has checkboxes to allow the user to pick items from the available
options.

CCIBI-4877
