+++ src/MultiSelectCheckbox/index.ts
@@ -1,0 +1,5 @@
+import { MultiSelectCheckbox } from './MultiSelectCheckbox.component';
+
+export { MultiSelectCheckbox } from './MultiSelectCheckbox.component';
+
+export default MultiSelectCheckbox;
